Use explicitly sized types from stdint
Replace bare int types with stdint types. This was done with:

    grep -rwl 'int' src/ | xargs sed -i 's/\<int\>/int16_t/g'
    grep -rwl 'unsigned long' src/ | xargs sed -i 's/\<unsigned long\>/uint32_t/g'
    grep -rwl 'unsigned char' src/ | xargs sed -i 's/\<unsigned char\>/uint8_t/g'

Then reverted for *main(), putchar(), and getchar().

The Arduino declarations for parallel_main() were also corrected to
match their definitions.

SDCC does *not* generate the same code in all instances, due to `int`
being treated different than `short int`.

+++ src/board/arduino/mega2560/battery.c
@@ -4,17 +4,17 @@
 
 #include <common/i2c.h>
 
-int smbus_read(uint8_t address, uint8_t command, uint16_t * data) {
+int16_t smbus_read(uint8_t address, uint8_t command, uint16_t * data) {
     return i2c_get(NULL, address, command, (uint8_t *)data, 2);
 }
 
-int smbus_write(uint8_t address, uint8_t command, uint16_t data) {
+int16_t smbus_write(uint8_t address, uint8_t command, uint16_t data) {
     return i2c_set(NULL, address, command, (uint8_t *)&data, 2);
 }
 
 void battery_debug(void) {
     uint16_t data = 0;
-    int res = 0;
+    int16_t res = 0;
 
     #define command(N, A, V) { \
         printf(#N ": "); \
